fix how product variant name is set for configurable product
