+++ ruleset/phpstan-baseline.neon
@@ -657,7 +657,7 @@ parameters:
 
 		-
 			message: "#^Cannot access offset 'parent' on mixed\\.$#"
-			count: 1
+			count: 3
 			path: ../src/Task/Product/BatchProductsTask.php
 
 		-
