+++ src/Processor/ProductVariant/PricingProcessorInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Synolia\SyliusAkeneoPlugin\Processor\ProductVariant;
+
+interface PricingProcessorInterface extends ProductVariantProcessorInterface
+{
+}
